+++ src/controller/all/all.go
@@ -54,6 +54,7 @@ import (
 	_ "github.com/faradey/madock/v4/src/controller/bigcommerce"
 	_ "github.com/faradey/madock/v4/src/controller/bigcommerce/setup"
 	_ "github.com/faradey/madock/v4/src/controller/custom/setup"
+	_ "github.com/faradey/madock/v4/src/controller/general/project/orphans"
 	_ "github.com/faradey/madock/v4/src/controller/general/snapshot/create"
 	_ "github.com/faradey/madock/v4/src/controller/general/snapshot/restore"
 	_ "github.com/faradey/madock/v4/src/controller/general/ssl"