+++ src/helper/cli/arg_struct/args.go
@@ -219,6 +219,10 @@ type ControllerGeneralServiceList struct {
 	attr.Arguments
 }
 
+type ControllerGeneralProjectOrphans struct {
+	attr.Arguments
+}
+
 type ControllerGeneralDbExecute struct {
 	attr.Arguments
 	DBServiceName string `arg:"-s,--service" help:"DB service name. For example: db"`